feat: oxapay charge

+++ src/main/kotlin/ltd/guimc/web/altget/entity/db/coin/UserCoin.kt
@@ -11,7 +11,7 @@ class UserCoin {
     @TableId
     var userId: Int = 0
 
-    var balance: Int = 0
+    var balance: Long = 0
 
     @TableField(fill = FieldFill.INSERT)
     var createdAt: LocalDateTime? = null
